+++ aimmo-game/tests/functional/test_invulnerability_pickups_and_effects.py
@@ -25,6 +25,14 @@ class TestInvulnerabilityPickupsAndEffects(TestCase):
         """
         self.cell.pickup = InvulnerabilityPickup(self.cell)
         self.assertEqual(self.game.avatar_manager.get_avatar(1).resistance, 0)
+        self.assertEqual(self.cell.pickup.serialise(), {
+            'type': 'invulnerability',
+            'location': {
+                'x': 0,
+                'y': 0,
+            }
+        })
+
 
         loop = asyncio.get_event_loop()
         loop.run_until_complete(self.game.simulation_runner.run_single_turn(self.game.avatar_manager.get_player_id_to_serialised_action()))
